Add client connect (unix only)
In current implementation Client must be created from RawFd handle.
This forces crate users to import nix dependency in their projects just to create a Client instance.
So this PR adds a shortcut Client::connect(path) to make it more user friendly.
