+++ example/async-client.rs
@@ -5,28 +5,13 @@
 
 mod protocols;
 
-use nix::sys::socket::*;
 use protocols::r#async::{agent, agent_ttrpc, health, health_ttrpc};
 use ttrpc::context::{self, Context};
 use ttrpc::r#async::Client;
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() {
-    let path = "/tmp/1";
-
-    let fd = socket(
-        AddressFamily::Unix,
-        SockType::Stream,
-        SockFlag::empty(),
-        None,
-    )
-    .unwrap();
-    let sockaddr = path.to_owned() + &"\x00".to_string();
-    let sockaddr = UnixAddr::new_abstract(sockaddr.as_bytes()).unwrap();
-    let sockaddr = SockAddr::Unix(sockaddr);
-    connect(fd, &sockaddr).unwrap();
-
-    let c = Client::new(fd);
+    let c = Client::connect_unix("/tmp/1").unwrap();
     let mut hc = health_ttrpc::HealthClient::new(c.clone());
     let mut ac = agent_ttrpc::AgentServiceClient::new(c);
